+++ invisible_cities/io/hits_io_test.py
@@ -5,7 +5,7 @@ import pandas as pd
 from . dst_io              import load_dst
 from numpy.testing import assert_allclose
 
-from .. core.test_utils    import assert_dataframes_equal
+from .. core.testing_utils import assert_dataframes_equal
 from ..types.ic_types      import xy
 from ..evm.event_model     import Cluster
 from ..evm.event_model     import Hit
